Split browser action routing helpers
